Rename graphdata file
