+++ tierkreis/tests/controller/test_resume.py
@@ -5,7 +5,7 @@ from uuid import UUID
 
 import pytest
 
-from tests.controller.eval_graphdata import double_partial
+from tests.controller.partial_graphdata import double_partial
 from tests.controller.sample_graphdata import (
     map_with_str_keys,
     maps_in_series,
